Version 0.3.1
Major edits to the ANOVA code.

NOTES:

1. Newly supported ANOVA models:
— anova3rm  (three-way design with repeated-measures on all three factors)
- anova2onerm  (now supports unbalanced designs:  i.e. different numbers of subjects for each level of factor A)


2. Non-sphericity corrections
— Now only available for two-sample t test and one-way ANOVA.
— The correction for one-way ANOVA is approximate and has not been validated.
— Non-sphericity corrections for other designs are currently being checked.


3. Repeated-measures ANOVA:
IF (a) the data are 1D and (b) there is only one observation per subject and per condition...
THEN inference is approximate, based on approximated residuals.
TO AVOID THIS PROBLEM:  use multiple observations per subject per condition, and the same number of observations across all subjects and conditions.